amp-class {TIMP}	R Documentation

Class "amp" for diagonal matrix model specification.

Description

amp is the class for diagonal matrix model specification; such models are internally initialized when a tri-linear-type model is fit to the data via passing the argument opt to fitModel as an object of class opt in which the slot trilinear has the value TRUE. All objects of class amp are sub-classes of class dat; see documentation for dat for a description of these slots.

Details

See kin-class for an example of the initialization of a kin object via the initModel function.

Objects from the Class

Objects can be created by calls of the form new("amp", ...) or amp(...).

Slots

amp

list of vectors of starting values for the parameters of the amplitudes for each dataset; one vector of values is used to parameterize the values corresponding to each dataset.

Extends

Class kin-class, directly.

Author(s)

Katharine M. Mullen, Ivo H. M. van Stokkum

See Also

kin-class, spec-class, opt-class

[Package TIMP version 1.5 Index]