baseIRF {TIMP}	R Documentation

Baseline subtraction from a vector, usually representing an IRF.

Description

Baseline subtraction from a vector, usually representing an IRF.

Usage

baseIRF(irfvec, indexlow, indexhigh, removeNeg = FALSE) 

Arguments

irfvec 	 Vector to subtract a baseline from

indexlow 	 Lowest index to base the baseline estimation on

indexhigh 	 Highest index to base the baseline estimation on

removeNeg 	 Whether negative values should be replaced with 0.

Details

Currently estimates the baseline as the mean of data between indexlow and indexhigh, and subtracts the result from the entire vector.

Value

vector

Author(s)

Katharine M. Mullen, Ivo H. M. van Stokkum

Examples

 
irfvec <- rnorm(128, mean=1) 
plot(irfvec,type="l") 
irfvec_corrected <- baseIRF(irfvec, 1, 10)
lines(irfvec_corrected, col=2)

[Package TIMP version 1.5 Index]