examineFit {TIMP}	R Documentation

Examines the results of a call to fitModel

Description

Examine the results of a call to fitModel by a call to plotting functions; call this function with argument an object returned from fitModel. Possibly also supply a new specification of plots to be generated.

Usage

examineFit(resultfitModel, opt=vector())

Arguments

resultfitModel 	 list returned by a call to fitModel

opt 	 possibly an object of class opt giving options for plotting; if opt has length zero (the default) then the plotting options given in the opt list of resultFitModel are applied

Details

The fitModel function returns a list of results, and initiates plotting functions. Given the resultfitModel list fitModel returns, examineFit initiates the plotting functions, and thus may be used to examine results.

Author(s)

Katharine M. Mullen, Ivo H. M. van Stokkum

See Also

fitModel, opt

[Package TIMP version 1.5 Index]