fit-class {TIMP}	R Documentation

Class "fit" to store the results of model fitting associated with all datasets analyzed.

Description

Class to store results of model fitting associated with all datasets in a single call to the fitModel function. An object of class fit is stored in the slot fit of objects of class multimodel.

Objects from the Class

Objects can be created by calls of the form new("fit", ...).

Slots

resultlist:

Object of class "list" that contains an object of class res for each dataset modeled, in the order that they were specified.

nlsres:

Object of class "list" containing named elements

onls

output of the call to nls used in model optimization.

sumonls

result of call summary(onls)

Author(s)

Katharine M. Mullen, Ivo H. M. van Stokkum

See Also

res-class, multimodel-class

[Package TIMP version 1.5 Index]