getClpindepX-methods {TIMP}	R Documentation

Generic function getClpindepX in Package ‘TIMP’

Description

Gets the matrix associated with nonlinear parameter estimates for the case that this matrix is not re-calculated per conditionally linear parameter.

Usage

getClpindepX(model, multimodel, theta, returnX, rawtheta, dind)

Arguments

model 	 Object of class dat; function switches on this argument.

multimodel 	 Object of class multimodel used in standard error determination

theta 	 Vector of nonlinear parameter estimates.

returnX 	 logical indicating whether to return a vectorized version of the X matrix

rawtheta 	 vector of nonlinear parmeters; used in standard error determination

dind 	 numeric indicating the dataset index; used in standard error determination

Author(s)

Katharine M. Mullen, Ivo H. M. van Stokkum

See Also

dat-class

[Package TIMP version 1.5 Index]