getResid {TIMP}	R Documentation

For data correction, fits a model (but ignores plotting commands) in order to obtain the SVD of the residuals, which then can be used in data-correction.

Description

For data correction, fits a model exactly as does fitModel (but ignores plotting commands) in order to obtain the SVD of the residuals. These residuals can then be subtracted away from the original data to some extent with the preProcess function.

Usage

getResid(data, modspec=list(), datasetind = vector(), modeldiffs = list(), 
                opt = opt() )

Arguments

data 	 As in the fitModel function

modspec 	 As in the fitModel function

datasetind 	 As in the fitModel function

modeldiffs 	 As in the fitModel function

opt 	 As in the fitModel function

Value

list containing the first five left and right singular vectors of the residuals, as well as the first five singular values. A weight matrix (if used) is also included in this list.

See Also

fitModel, preProcess

[Package TIMP version 1.2 Index]