getResults {TIMP}	R Documentation

Functions to print and return parts of the object returned by the fitting routines.

Description

Functions to print and return parts of the object returned by fitModel. onls returns the output of the nls function. sumonls returns the result of calling summary on onls function. parEst returns a summary of model parameter estimates. The remaining functions return lists representing various aspects of the results returned by the function fitModel.

Usage

onls(result)
sumnls(result) 
parEst(result, param = "", dataset = NA, verbose = TRUE)
getXList(result)
getCLPList(result, getclperr = FALSE) 
getDataList(result, numsing = 2)
getResidualList(result, numsing = 2)
getTracesList(result)
getdim1List(result) 
getdim2List(result)

Arguments

result 	 return value of fitModel

param 	 character vector of the particular parameters to return; if param=="" then all parameters are given.

dataset 	 index of the dataset from which to return results; by default dataset=NA in which case results from all datasets are returned

verbose 	 logical that defaults to TRUE that determines whether parEst just returns a list invisibly or prints as well.

getclperr 	 logical that defaults to FALSE that determines whether a list containing the standard error estimates associated with the conditionally linear parameters, as opposed to the conditionally linear parameters themselves

numsing 	 integer that defaults to 2; determines the number of singular vectors to return

Value

sumnls returns an object of class "summary.nls".
onls returns an object of class "nls".
parEst returns an object of class "list" representing the parameter estimates.
XList returns a "list" of length equal to the number of datasets modeled, where each element represents the matrix determined by the nonlinear parameters (under a kinetic model, the concentrations).
CLPList returns a "list" of length equal to the number of datasets modeled, where each element represents the matrix determined as conditionally linear parameters (under a kinetic model, the spectra).
getDataList returns a "list" of length 2. The first element is named svddatalist and is a "list" of length equal to the number of datasets modeled, where each element is a list with elements values, left and right, where values contains the singular values, left contains numsing left singular vectors, and right contains numsing right singular vectors, all of the residuals. The second element is named "datalist" and is a "list" of length equal to the number of datasets modeled, where each element represents the matrix of unweighted data.
getResidualList returns a "list" of length 2. The first element is named svdresidlist and is a "list" of length equal to the number of datasets modeled, where each element is a list with elements values, left and right, where values contains the singular values, left contains numsing left singular vectors, and right contains numsing right singular vectors, all of the residuals. The second element is named "residlist" and is a "list" of length equal to the number of datasets modeled, where each element represents the matrix of residuals.
getTracesList returns a "list" of length equal to the number of datasets modeled, where each element is a "list" of length equal to the number of x2 values in the dataset (wavelengths for kinetic models). Each of the x2 elements represents the trace, that is, fit with respect to time, at the corresponding x2-value.
getdim1List returns a "list" of length equal to the number of datasets modeled, where each element is a "vector" of length equal to the number of x values in the dataset (times for kinetic models).
getdim2List returns a "list" of length equal to the number of datasets modeled, where each element is a "vector" of length equal to the number of x2 values in the dataset (wavelengths for kinetic models).

Author(s)

Katharine M. Mullen, Ivo H. M. van Stokkum

See Also

fitModel

[Package TIMP version 1.5 Index]