kin-class {TIMP}	R Documentation

Class "kin" for kinetic model storage.

Description

kin is the class for kinetic models; an object of class "kin" is initialized if mod_type = "kin" is an argument of initModel. All objects of class kin are sub-classes of class dat; see documentation for dat for a description of these slots.

Details

See dat-class for an example of the initialization of a kin object via the initModel function.

Objects from the Class

Objects can be created by calls of the form new("kin", ...) or kin(...).

Slots

kinpar

vector of rate constants to be used as starting values for the exponential decay of components; the length of this vector determines the number of components of the kinetic model.

specpar:

Object of class "list" parameters for spectral constraints

seqmod:

Object of class "logical" that is TRUE if a sequential model is to be applied and FALSE otherwise

irf:

Object of class "logical" that is TRUE is an IRF is modeled and FALSE otherwise

mirf:

Object of class "logical" that is TRUE if a measured IRF is modeled and FALSE otherwise

measured_irf:

Object of class "vector" containing a measured IRF

convalg:

Object of class "numeric" 1-6 determining the numerical convolution algorithm used in the case of modeling a measured IRF; if 6 then supply a reference lifetime in the slot reftau.

reftau:

Object of class "numeric" containing a reference lifetime to be used when convalg=6

irffun:

Object of class "character" describing the function to use to describe the IRF, by default "gaus"

irfpar:

Object of class "vector" of IRF parameters; for the common Gaussian IRF this vector is ordered c(location, width)

dispmu:

Object of class "logical" that is TRUE if dispersion of the parameter for IRF location is to be modeled and FALSE otherwise

dispmufun:

Object of class "character" describing the functional form of the dispersion of the IRF location parameter; if equal to "discrete" then the IRF location is shifted per element of x2 and parmu should have the same length as x2. defaults to a polynomial description

parmu:

Object of class "list" of starting values for the dispersion model for the IRF location

disptau:

Object of class "logical" that is TRUE if dispersion of the parameter for IRF width is to be modeled and FALSE otherwise

disptaufun:

Object of class "character" describing the functional form of the dispersion of the IRF width parameter; if equal to "discrete" then the IRF width is parameterized per element of x2 and partau should have the same length as x2. defaults to a polynomial description

partau:

Object of class "vector" of starting values for the dispersion model for the IRF FWHM

fullk:

Object of class "logical" that is TRUE if the data are to be modeled using a compartmental model defined in a K matrix and FALSE otherwise

kmat:

Object of class "array" containing the K matrix descriptive of a compartmental model

jvec:

Object of class "vector" containing the J vector descriptive of the inputs to a compartmental model

ncolc:

Object of class "vector" describing the number of columns of the C matrix for each clp in x2

kinscal:

Object of class "vector" of starting values for branching parameters in a compartmental model

kmatfit:

Object of class "array" of fitted values for a compartmental model

cohspec:

Object of class "list" describing the model for coherent artifact/scatter component(s) containing the element type and optionally the element numdatasets if type is "irf", the coherent artifact/scatter has the time profile of the IRF. if type is "freeirfdisp" the coherent artifact/scatter has a Gaussian time profile whose location and width are parameterized in the vector coh. if type is "irfmulti" the time profile of the IRF is used for the coherent artifact/scatter model, but the IRF parameters are taken per dataset (for the multidataset case), and the integer argument numdatasets must be equal to the number of datasets modeled. if type is "seq" a sequential exponential decay model is applied, whose parameters are contained in coh. if type is "mix" a sequential exponential decay model is applied along with a model that follows the time profile of the IRF; the coherent artifact/scatter is then a linear superposition of these two models.

coh:

Object of class "vector" of starting values for the parameterization of a coherent artifact

wavedep:

Object of class "logical" describing whether the kinetic model is dependent on x2 index (i.e., whether there is clp-dependence)

lambdac:

Object of class "numeric" for the center wavelength to be used in a polynomial description of x2-dependence

Extends

Class dat-class, directly.

Author(s)

Katharine M. Mullen, Ivo H. M. van Stokkum

See Also

dat-class, spec-class

[Package TIMP version 1.4 Index]