kinopt-class {TIMP}	R Documentation

Class "kinopt" stores options for fitting and plotting kinetic models

Description

Class "kinopt" stores options for fitting and plotting kinetic models in particular; this is a subclass of class opt

Details

See opt-class for the specification of fitting/plotting options that are not specific to the class type.

Objects from the Class

Objects can be created by calls of the form new("kinopt", ...) or kinopt(...)

Slots

notraces:

Object of class "logical" that defaults to FALSE; if TRUE, do not plot traces

selectedtraces:

Object of class "vector" containing x indices for which plots of traces are desired under a kinetic model

breakdown:

Object of class "list" with the following elements:

plot

vector of x2 values to plot the breakdown for. These values be specified in a fuzzy way: an x2 value within abs(x2[1] - x2[2])/100 a value given in plot means that a plot for that x2 value will be generated, where the reference x2[1] and x2[2] are from the first dataset modelled.

tol

numeric giving a tolerance by which the values in plot are compared to x2 values for near-equality. The default is defined as abs(x2[1] - x2[2])/100.

superimpose

vector of dataset indices for which results should be superimposed if the dataset has an x2 value at a value in plot.

FLIM:

Object of class "logical" that defaults to FALSE; if TRUE, the data represent a FLIM experiment and special plots are generated.

FLIMresidimag:

Object of class "logical" that defaults to TRUE; if FALSE and a FLIM image is analyzed, the residuals are not plotted as an image.

noFLIMsummary:

Object of class "logical" that defaults to FALSE; if TRUE and a FLIM image is analyzed, only other plots requested by the user (such as traces or residuals) are generated, and no summary plot in made.

kinspecest

Object of class "logical" that defaults to FALSE; if TRUE, make a plot of the spectra associated with the kinetic components as well as the lifetime estimates.

writeplaincon

Object of class "list"; if length is greater than 0, then the concentration model will be evaluated at the vector of x values supplied as the element "x" of writeplaincon and the result will be written to file for each dataset.

writerawcon

Object of class "logical" that defaults to FALSE; if TRUE, then the representation of the concentration profiles before the application of constraints (to account for the equality of spectra, etc.) is written to file for each dataset.

plotcohcolspec

Object of class "logical" that defaults to TRUE; if FALSE then the spectra associated with the coherent artifact (pulse-follower) are not included in the summary plots

Author(s)

Katharine M. Mullen, Ivo H. M. van Stokkum

See Also

examineFit, fitModel, opt-class, specopt-class

[Package TIMP version 1.4 Index]