mass-class {TIMP}	R Documentation

Class "mass" for mass spectrometry model storage.

Description

mass is the class for mass spectrometry models; an object of class "mass" is initialized if mod_type = "mass" is an argument of initModel. All objects of class mass are sub-classes of class kin; see documentation for kin for a description of these slots.

Details

See kin-class for an example of the initialization of a kin object via the initModel function.

Objects from the Class

Objects can be created by calls of the form new("mass", ...) or kin(...).

Slots

peakpar

list of vectors of starting values for the parameters of components; one vector of values is used to parameterize each component.

peakfunct:

Object of class "character" that specifies the function by which components are parameterized in time; this is by default "expmodgaus" for the exponentially modified Gaussian function.

lzerofile:

Object of class "character" that specifies the filename of the lzero specification to read in from file. This file has the format: 1st line not read; lines thereafter are the space-delimited index of the component to constrain, the lower bound of the constraint, and the upper bound of the constraint, e.g., 1 218.800000000000011 220.099999999999994

extracomp:

Object of class "logical" that defaults to TRUE and determines whether a component with constant concentration in time is added to the model to represent a baseline.

shift:

Object of class "vector" that represents a shift of the location of each elution profile peak; this can be specified per-component, in which case length(shift) is the number of components (not including a baseline component) or for all components, in which case length(shift == 1).

Extends

Class kin-class, directly.

Author(s)

Katharine M. Mullen, Ivo H. M. van Stokkum

See Also

kin-class, spec-class

[Package TIMP version 1.5 Index]