massopt-class {TIMP}	R Documentation

Class "massopt" stores options for fitting and plotting models for mass spectrometry data

Description

Class "massopt" stores options for fitting and plotting models models for mass spectrometry data in particular; this is a subclass of class opt that contains options applicable to all model types

Details

See opt-class and for the specification of fitting/plotting options that are not specific to the mass class type.

Objects from the Class

Objects can be created by calls of the form new("massopt", ...) or massopt(...)

Slots

axis.by:

Object of class "numeric" that allows labels on the bars representing the mass spectra to to skipped, e.g., axis.by=2 will add a label to every second bar

scale.concen:

Object of class "logical" that scales the concentration matrix using the algorithm found in the function scaleConList.

nummaxtraces:

Object of class "nummaxtraces" that defaults to zero; if greater than zero then this number of the traces with the maximum amplitude are plotted

Author(s)

Katharine M. Mullen, Ivo H. M. van Stokkum

See Also

examineFit, fitModel, opt-class, specopt-class

[Package TIMP version 1.5 Index]