modifyModel {TIMP}	R Documentation

Allows the starting values for parameters associated with a model to be updated with the values found in fitting the model.

Description

Allows the starting values for parameters associated with a model to be updated with the values found in fitting the model. A call model_w_new_starting_vals <- modifyModel(old_model) will plug in the optimized parameter values the last model fit so that are the starting values in the model specification model_w_new_starting_vals.

Usage

modifyModel(model = list(), newest = list(), exceptslots = vector() )

Arguments

model 	 an object of class dat returned by initModel; if this argument is of length(0), which is the default, then the last model fit is used (which is found in the global variable .currModel@model)

newest 	 an object of class theta containing new parameter estimates; if this argument is of length(0), which is the default, then the parameter estimates associated with dataset 1 in the last model fit are used (which are found in the global variable .currTheta[[1]])

exceptslots 	 a vector of character vector of slot names whose corresponding slots are to be left out of the update

Value

an object of class dat that returns the results of calling initModel with the new starting values.

Author(s)

Katharine M. Mullen, Ivo H. M. van Stokkum

See Also

initModel, fitModel

[Package TIMP version 1.2 Index]