multimodel-class {TIMP}	R Documentation

Class "multimodel" for storage of multidataset models, data and the results of fitting.

Description

multimodel is the class to store data, a generally applicable model, a list of per-data models, a specification of per-dataset model differences, and results for the analysis of possibly many datasets. After a call to fitModel an object .currModel is initialized of the multimodel class.

Details

after a call to fitModel, an object of class multimodel exists in the global environment as the variable .currModel

Objects from the Class

Objects can be created by calls of the form new("multimodel", ...) or multimodel(...).

Slots

data:

Object of class "list" of objects of class dat containing data

model:

Object of class "dat" of class dat containing a model specification to be applied to all datasets

modellist:

Object of class "list" of length n where n is the number of datasets given in data, and each element i is an object of class dat giving the dataset-specific model applicable to data[[i]]

modeldiffs:

Object of class "list" of per-dataset model differences input as an argument to the fitModel function

fit:

Object of class "fit" containing a list of results per-dataset as well as the output of optimization returned by the nls function.

groups:

Object of class "list" containing a list of lists of the groups of clp to link across datasets. Each component list contains vectors of form (clp condition index, dataset index), and such vectors in the same component list are linked between datasets. See fitModel for more details on the linking possibilities.

stderrclp:

Object of class "logical" describing whether standard error estimates on conditionally linear parameters should be calculated; this is determined by the opt argument of fitModel and defaults to FALSE

Author(s)

Katharine M. Mullen, Ivo H. M. van Stokkum

See Also

fitModel

[Package TIMP version 1.4 Index]