multitheta-class {TIMP}	R Documentation

Class "multitheta" that stores a list with one element of class "theta" for each dataset modeled.

Description

Class multitheta stores a list with one element of class theta for each dataset modeled, corresponding to the parameter estimates associated with that dataset.

Objects from the Class

Objects can be created by calls of the form new("multitheta", ...) or multitheta(...).

Slots

th:

Object of class "list" with element i corresponding to the theta object for the ith dataset modeled.

Author(s)

Katharine M. Mullen, Ivo H. M. van Stokkum

See Also

theta-class, dat-class

[Package TIMP version 1.5 Index]