opt-class {TIMP}	R Documentation

Class "opt" stores options for fitting and plotting

Description

Class "opt" stores options for fitting and plotting applicable to all model types

Details

See kinopt-class, specopt-class and massopt-class for the specification of fitting/plotting options that are specific to the class type.

Objects from the Class

Objects can be created by calls of the form new("opt", ...) or opt(...).

Slots

nls:

Object of class "logical" that defaults to TRUE; if nls=TRUE the Gauss-Newton algorithm implemented in the function nls is used for parameter optimization; if nls=FALSE the Levenberg-Marquardt algorithm ported from MINPACK by the nls.lm function of the package "minpack.lm" is used.

nnls:

Object of class "numeric" that defaults to 0. It has the following meaning:

0:

Do not constrain the conditionally linear parameters to nonnegativity

1:

Constrain the conditionally linear parameters to nonnegativity via an nonnegative least squares algorithm as implemented via the function optim with method="L-BFGS-B"

2:

Constrain the conditionally linear parameters to nonnegativity via an nonnegative least squares algorithm as implemented via the function nnls from the package by the same name.

3:

Constrain the conditionally linear parameters to nonnegativity via an nonnegative least squares algorithm as implemented via a quadratic program using the function solve.qp from the package "quadprog".

writecon:

Object of class "logical" that defaults to FALSE; if true then concentrations are written to a txt file; row labels are x

writespec:

Object of class "logical" that defaults to FALSE; if TRUE then spectra are written to a txt file; row labels are x2

writenormspec:

Object of class "logical" that defaults to FALSE; if TRUE then normalized spectra are written to a txt file; row labels are x2

writefit:

Object of class "logical" that defaults to FALSE; if TRUE then fit is written to a txt file; row and column labels are x and x2

writeclperr:

Object of class "logical" that defaults to FALSE; if true then the error bars for clp are written to a txt file. This option is only sensible with stderrclp=TRUE.

output:

Object of class "character" that defaults to "ps", which means that plots written to file are postscript. Alternatively, specify output = "pdf", and plots are written as pdf files

addfilename:

Object of class "logical" that, for each data file, tries to add the filename to plots associated with output for that data.

residplot:

Object of class "logical" defaults to FALSE; if TRUE generate a plot of residuals in a separate window.

plot:

Object of class "logical" that defaults to TRUE; if FALSE then do not write output in the form of plots and other windows to the screen.

divdrel:

Object of class "logical" that defaults to FALSE; if TRUE, plot traces and concentration profiles divided by the dataset scaling parameters where they apply; this allows for the fit of datasets having different intensities on the same scale.

plotkinspec:

Object of class "logical" that defaults to FALSE; if TRUE, generates a separate plot of the spectra associated with the components that are not a part of a coherent artifact/scatter model.

superimpose:

Object of class "vector" containing dataset indices whose results should be superimposed in plots

xlab:

Object of class "character" containing label for x-axis, e.g., "nanoseconds" or "picoseconds"

ylab:

Object of class "character" containing label for y-axis, e.g., "wavelength"

title:

Object of class "character" containing title to write at the top of plots.

makeps:

Object of class "character" containing prefix to plot files written to postscript; if present postscript will be written. Note that this string is also used as the preffix of txt output files

linrange:

Object of class "numeric" giving linear range of time axis for plotting; time will be plotted linearly from -linrange to linrange and plotted on a logarithmic (base 10) axis elsewhere

summaryplotrow:

Object of class "numeric" giving number of rows in summary plot; defaults to 4

summaryplotcol:

Object of class "numeric" giving number of columns in summary plot; defaults to 4

iter:

Object of class "numeric" giving number of iterations to optimize model parameters

paropt:

Object of class "list" of graphical parameters in format par(...) to apply to plots.

stderrclp:

Object of class "logical" that defaults to FALSE; if TRUE, estimates of the standard error of conditionally linear parameters are made

addest:

Object of class "vector" containing character strings of which parameter estimates should be added to the summary plot, e.g., addest = c("kinpar", "irfpar")

kinspecerr

Object of class "logical" that defaults to FALSE; if TRUE, add standard error estimates to the clp a plot generated with kinspecest=TRUE or plotkinspec=TRUE. This option can only be used if the estimates were generated during fitting via the option stderrclp=TRUE

xlimspec

Object of class "vector" that defaults to vector(); if changed, it should specify the desired x-limits of the plot of clp

ylimspec

Object of class "vector" that defaults to vector(); if changed, it should specify the desired y-limits of the plot of clp. In the case of plotting the results of FLIM image analysis, ylimspec can be used to determine the range used in the image plot of lifetimes.

ylimspecplus

Object of class "vector" that defaults to vector(); if changed, the first value should specify a vector to add to the y-limits of the plot of clp

samespecline

Object of class "logical" that defaults to FALSE; if TRUE, then the line-type for clp is the same for all datasets

specinterpol

Object of class "logical" that defaults to FALSE; if TRUE, use spline instead of lines between the points representing estimated clp

specinterpolpoints

Object of class "logical" that defaults to TRUE; if TRUE, add points representing the actual estimates for clp to plots of the curves respresenting smoothed clp

specinterpolseg

Object of class "numeric" that defaults to 50; represents the number of segments used in a spline-based representation of clp

specinterpolbspline

Object of class "logical" that defaults to FALSE; determines whether a B-spline based representation of clp is used (when specinterpol=TRUE) or a piecewise polynomial representation

normspec

Object of class "logical" that determines whether clp are normalized in plots

writespecinterpol

Object of class "logical" that defaults to FALSE; if TRUE, a spline-based representation of clp is written to ASCII files

nlsalgorithm

Object of class "character" that defaults to "default" and determines the algorithm used by nls, if nls is used in optimization. See help(nls) for other possibilities, such as "port", which is more stable with respect to starting values but requires more time.

Author(s)

Katharine M. Mullen, Ivo H. M. van Stokkum

See Also

kinopt-class, specopt-class

[Package TIMP version 1.5 Index]