outlierCorr {TIMP}	R Documentation

Finds and removes outliers from a datasets

Description

Finds and removes outliers from datasets given the results of fitting as returned by fitModel. Uses the residuals in the fitted results to return a list of corrected datasets to be used in place of the datasets used in the call to fitModel as well as a list of weights. The data returned contains the fitted values at pointed that are outliers and will be assigned zero weight in subsequent fits.

Usage

outlierCorr(oldRes, thresh=.8) 

Arguments

oldRes 	 Object returned by fitModel function

thresh 	 Object of class "numeric" determining what points to consider outliers.

Details

The points that are outliers are those points xx at a given value j of x2 that satisfy thresh < resid[j,]/max(fitted[j,xx], max(fitted[j,xx])/10) . Increasing thresh will assign fewer points as outliers.

Value

list containing the elements dt, a list of corrected datasets, and weightList, a list of new weight matrices.

See Also

fitModel, preProcess

[Package TIMP version 1.5 Index]