readclp0 {TIMP}	R Documentation

This function reads in a specification of constraints to zero on the clp.

Description

This function is useful for the case that there are many constraints to zero in the model, as is the case for some mass spectrometry models.

Usage

readclp0(filenm)

Arguments

filenm 	 Object of class "character" that gives is the path to the file to read in.

Details

The file to be read in should have the following format: 1st line is not read. Lines thereafter are the space-delimited index of the component to constrain, the lower bound of the constraint, and the upper bound of the constraint, e.g., 1 218.800000000000011 220.099999999999994.

Value

The constraints to zero in the format documented in the help file for the "dat" class. Therefore a call to "readclp0" may be used inside a call to "initModel", as in clp0 = readclp0("filename").

Author(s)

Katharine M. Mullen, Ivo H. M. van Stokkum

See Also

initModel

[Package TIMP version 1.5 Index]