res-class {TIMP}	R Documentation

Class "res" to store the results of model fitting associated with a single dataset.

Description

Class to store results of model fitting associated with a single dataset. A list containing objects of class res is a slot in class fit. An object of class fit is stored in the slot fit of objects of class multimodel.

Objects from the Class

Objects can be created by calls of the form new("res", ...). A res object is created after model fitting via the residual function residPart.

Slots

cp:

Object of class "list" that contains the estimates for conditionally linear parameters.

resid:

Object of class "list" of residuals, with one element for each dataset modeled.

fitted:

Object of class "list" of fits, with one element for each dataset modeled.

irfvec:

Object of class "list" with a vector of elements for each element of the clp x2

Author(s)

Katharine M. Mullen, Ivo H. M. van Stokkum

See Also

fit-class, multimodel-class

[Package TIMP version 1.2 Index]