residPart-methods {TIMP}	R Documentation

Generic function residPart in Package ‘TIMP’

Description

Methods for function residPart in Package `TIMP' determine the part of the residual vector associated with a single `part' of the dataset(s).

Usage

residPart(model, group, multimodel, thetalist, clpindepX, finished, returnX,
rawtheta)

Arguments

model 	 Object of class dat; switches on this argument.

group 	 list of vector pairs (x2 index, dataset index) for which the part of the residual vector is to be determined

multimodel 	 Object of class multimodel

thetalist 	 Object of class multitheta

clpindepX 	 Object of class matrix containing the matrix determined directly by the nonlinear parameters (e.g., a concentration matrix in the case of a kinetic model) in the case that this matrix does not depend on the x2 index

finished 	 logical determining whether fitting is finished that triggers the storage of results

returnX 	 logical determining whether to just return the matrix X directly dependent on nonlinear parameters; this is used in the finite difference derivative of X used to get standard error estimates on the conditionally linear parameters.

rawtheta 	 numeric vector of nonlinear parameters to be optimized by nls; this is used in the finite difference derivative of X used to get standard error estimates on the conditionally linear parameters.

See Also

dat-class, spec-class, kin-class

[Package TIMP version 1.5 Index]