specopt-class {TIMP}	R Documentation

Class "specopt" stores options for fitting and plotting spectral models

Description

Class "specopt" stores options for fitting and plotting spectral models in particular; this is a subclass of class opt.

Details

See opt-class for the specification of fitting/plotting options that are not specific to the class type.

Objects from the Class

Objects can be created by calls of the form new("specopt", ...). or specopt(...)

Slots

nospectra:

Object of class "logical" that defaults to FALSE; if TRUE, do not plot time-resolved spectra

selectedspectra:

Object of class "vector" containing x indices for which plots of time-resolved spectra are desired under a spectral model

Author(s)

Katharine M. Mullen, Ivo H. M. van Stokkum

See Also

opt-class, kinopt-class

[Package TIMP version 1.4 Index]