sumKinSpecEst {TIMP}	R Documentation

Makes a summary plot of spectra associated with kinetic components alongside a plot showing parameter estimates

Description

Makes a summary plot of spectra associated with kinetic components alongside a plot showing parameter estimates for, by default, kinetic parameters. If the analysis had more parameters in the addEst slot of the arguement opt, then more parameters are displayed. Note that this summary leaves out the spectra associated with coherent artifact or scatter.

Usage

sumKinSpecEst(listFits, addtitle = TRUE, customtitle = "", preps = "", 
ylimlist=list(), kinspecerr=TRUE)

Arguments

listFits 	 list of objects returned by the fitModel function

addtitle 	 logical regarding whether to add a title; if TRUE and customtitle is not given then the title is "Summary of EADS for: " plus the analysis titles

customtitle 	 character vector containing a title

preps 	 character vector describing the prefix of the postscript filename given as output

ylimlist 	 list with elements list(ind, ylim). ind is an index into listFits; ylim is the desired ylim for the plot for that analysis

kinspecerr 	 logical regarding whether to add error bars for to the estimated spectra.

Details

This looks best with less than five objects in listFits.

Value

Author(s)

Katharine M. Mullen, Ivo H. M. van Stokkum

See Also

fitModel, examineFit

[Package TIMP version 1.5 Index]