theta-class {TIMP}	R Documentation

Class "theta" for storage of nonlinear parameter estimates

Description

theta is the class to store parameter estimates associated with possibly many datasets; after a call to fitModel a list containing theta objects for each of the n datasets analyzed in the call to fitModel is created. To see the parameter estimates associated with the datasets, examine the object currTheta in the list returned by fitModel

Details

after a call to fitModel, an object of class theta exists in the global environment as the variable currTheta

Objects from the Class

Objects can be created by calls of the form new("theta", ...) or theta(...).

Slots

kinpar:

Object of class "vector" of rate constant estimates

specpar:

Object of class "list" of spectral shape parameter estimates

irfpar:

Object of class "vector" of IRF parameter estimates

parmu:

Object of class "list" of parameter estimates describing dispersion of the location of other parameters (in time, temp., etc.)

partau:

Object of class "vector" of parameter estimates describing dispersion of the width of other parameters (in time)

clpequ:

Object of class "vector" of parameter estimates describing conditionally linear parameters (spectra, in a kinetic model) relations

specdisppar:

Object of class "list" of parameter estimates describing dispersion of spectra

kinscal:

Object of class "vector" of parameters describing kinetic relations in the context of a compartmental scheme

prel:

Object of class "vector" of parameters describing relations between parameters (which may be linear, exponential, etc.)

coh:

Object of class "vector" of parameters describing a coherent artifact or pulse follower.

drel:

Object of class "vector" of parameters describing relations between datasets (linear, and possibly per-wavelength or, in general, per-clp)

Author(s)

Katharine M. Mullen, Ivo H. M. van Stokkum

See Also

fitModel , multitheta-class

[Package TIMP version 1.4 Index]